Create bubbles input consumer before nav bar long press
NavHandleLongPressInputConsumer handles long presses on the nav handle
and passes the touches on to a delegate if the touch is not a long
press.
When bubbles are expanded, we create a new input consumer that does
nothing and allows bubble input consumer in shell to take over.
Change the order how the input consumers are created. Create the bubbles
input consumer first, so it can be passed in as a delegate to the
nav handle input consumer. And this avoids bubbles overriding the nav
handle behavior.

Flag: ACONFIG com.android.wm.shell.enable_bubbles_long_press_nav_handle DISABLED
Bug: 324910035
Test: manual, expand a bubble, long press on home button for CtS
Change-Id: I05c219145c096a567bb0bb5a9cb3a40046b29519