Remove Unused parts of LauncherPrefs causing cyclical dependency.
This "migration" feature was planned as an effort to reduce latency upon
startup. That being said, now that there is a loading screen being added
to pixel startup time, it doesn't make sense to perform a risky data
migration just to save a 100ms or so of startup time. Therefore, I am
removing this code as a clean up effort to make LauncherPrefs simpler.
Also, there was a cyclical dependency introduced in b/324670265 that
makes this cleanup effort especially worthwhile, as it resolves a bug
for other developers.

Bug: 324670265
Test: Installed this on a device and tested that nothing was broken.
Flag: NA
Change-Id: I81df710057d80fbd071bc26a53cb191d453e7dee