Fix Launcher preview crash
Launcher preview could crash when unfolding a foldable device due to the
folded state display id no longer being valid, resulting in getting a
null display. This was leading to an empty preview bundle being
returned, which is unexpected because null is expected in the case of an
exception. Instead of crashing in getDisplayId, it's better to throw an
exception earlier on if the display is null and return null.

Flag: NA
Bug: 326637851
Test: manually verified by testing unfold in WPP in old and new preview

Change-Id: Ibc353d9447a71d33092d96029a812017487d0970