Fix issue that sometimes the widget list table occupies more space.
Since the widget cells are is recycled, some times, the cell occupies
more space based on the previous cell that it was recycled from. So,
we request layout to update it.

The issue is easily reproducible by first expanding app that has
single item but with large spanY. Then, expanding app with one item
but with smaller spanY (e.g. settings or sheets apps).

* Before: http://screencast/cast/NDkwOTMwOTcyMzkzNDcyMHxjMDdmZDE1Yi05Yg
* After: http://screencast/cast/NTA0NzU3Nzc5MDY0NDIyNHw1NzhkZmM0MS04OA

Bug: 289147379
Flag: N/A
Test: Manual
Change-Id: I4d003024e1db437a50de759f9390190b038a114a