Remove cyclical dependency from LauncherPrefs.
ag/26260351 was a much larger CL that removed more pieces of
LauncherPrefs that were related to this flag and were unused.
Unfortunately, that CL caused a test regression, and so I am publishing
this much smaller change in an effort to resolve the underlying issue in
a timely manner.

Bug: 324670265
Test: Installed this on a device and tested that nothing was broken.
Flag: NA
Change-Id: I10be75a1278fdffe7cc01beecf37b967526c4bc4