Remove line between bubble bar and arrow
Bubble bar background is a drawable that consists of a rounded rect and
a pointer arrow.
In light theme, the rounded rect has a 1 pixel wide border around it.
When drawing the arrow just above the rounded rect, the border is
visible as a line between the background rounded rect and the arrow.
Shifting the arrow down by 1 pixel to ensure that there is no visible
arrow between bubble bar and arrow.

Other option could be to draw the entire background as a single path.
This path can combine the rounded rect background and arrow. But we need
to control the alpha of the arrow separately from the background. This
we can't do with a single path.

Bug: 324442574
Flag: NA
Test: manual
Change-Id: Iaa58ae4700ed9ad43051cd386f03381e0c14e72c