Fix overview live tile flickers when clicking on overview action buttons
When a Live Tile is present on overview and an action is triggered (e.g., Split, Select, Screenshot), the live tile flickers due to a competing snapshot modification and animation on UI thread. The live tile is updated to switch to screenshot and at the same time the finishRecentsAnimation is triggered.

This CL bumps the number of frames to wait after the view is updated to run the Runnable code. Bumping back the mDeferFrameCount to 2 solves the sync between rendering the snapshot and hiding the Live Tile.

Fix: 327380570
Flag: N/A
Test: Manual. Open an App -> Swipe to Overview (Live title) -> Click on Split.
Change-Id: Ibd22a58580b2e5e5b866deb4e7dc87945c7e38c4