Add preview container size helper to display previews in picker.
While we want previews to be displayed at true size, today, there is
less consistency among different widgets. This change is a first step
towards consistent sizes. In this change, we provide helper classes
that provide the size of the container in terms of grid spans.

Note: This does not change the size for rendering widget previews; but,
we take that original rendered preview and scale it maintaining the
aspect ratio to display it in one of closest container sizes.

Bug: 319152349
Flag: N/A
Test: Includes a unit test & manual with the child cls
Change-Id: I335373aa1be9a41fe039c98cded0113a007ad8c4