Fix incorrect fragment in gesture tutorial when rotating the screen
This change makes the current fragment load on configuration changes for
the gesture navigation tutorial. Previously, the old fragment would be recreated causing inconsistencies when rotating the screen. Not marking a flag as this change is for both the previous gesture tutorial and the current one.

Fix: 317162126
Test: Run the tutorial and rotate the screen after each step. The
tutorial should progress like normal.
Flag: N/A

Change-Id: I09c40006dd0ef6e9e78fa44e36962a678918420a