Support going home from immersive fullscreen in 3 button nav
In gesture nav the swipe happens at the bottom of the screen,
which brings the navbar handle into the view.

In 3 button nav the swipe needs to happen from the right edge of
the screen, because the 3 button nav is always staying fixed in
the "bottom portrait" position, unless it is large screen, in
which case it should happen on the bottom landscape edge.

This can be detected via config_navBarCanMove resource value.

Bug: 313928629
Test: atest NetflixEnterPipTest
Test: above command on foldable in folded and unfolded state
Flag: n/a

Change-Id: I0a94757687fd9eb491b2a5cf0086ae51589ae0aa