Fix launcher activity leaking when desktop windowing enabled
When desktop windowing is enabled we initiate the
`SplitFromDesktopController` within the `SplitSelectStateController`
which registers a `SplitSelectListener` on the launcher instance.
However when we destroy the `SplitSelectStateController` the
`SplitFromDesktopController` is not destroyed an thus any listeners are
not unregisterd. Instead we should explicitly destroy the
`SplitFromDesktopController` by unregistering listener.

Flag: NONE
Fixes: 332667403
Bug: 331774319
Test: atest -c NexusLauncherTests:com.android.quickstep.TaplStartLauncherViaGestureTests
atest -c NexusLauncherTests:com.android.quickstep.util.SplitSelectStateControllerTest

Change-Id: I68ffcc4114644e75f751632eca8bc73e406139a8