Fix homescreen rotation delay during fold
The issue is caused by the delay that Launcher handles the rotation; onDisplayInfoChanged that RotationHelper listens to is triggered later than the screen turning on. onDeviceProfileChanged is triggered earlier, but it only receives callback when in foreground. Using both listeners means that if in the background, we can still rely on onDisplayInfoChanged to update, assuming by the delay is tolerable given it takes time to changes to foreground.

Bug: 300929233
Flag: NA
Test: Manual and Perfetto trace in the bug
Change-Id: I5c6983649a38c3e4fe2b8c4a950435ceb07afa5a