Show add button when WidgetCell text is tapped
This change adds the click listener to the cell itself as well as
the preview, so that clicks anywhere in the cell will show the add
button. Only long clicks on the preview image will start the
drag-and-drop (current behavior).

Bug: 323886237
Test: manual tap on Widget cell
Flag: ACONFIG com.android.launcher3.enable_widget_tap_to_add TEAMFOOD
Change-Id: I23b9b1261979971deffcf1132d95bd6a03c92c47