[omni] Log LP duration & slop's data source and value.
For duration, the data source and value should be the same for LPNH and LPH. E.g., if AGSA overrides duration, it applies for both LPNH and LPH. When there is no AGSA override, there are cases that one of our feature flag is enabled and the other one is disabled. We just simply rely on LPNH's feature flag in this case. This should be fine because the purpose is to cross check the data source and its value.

Bug: 332722082
Test: manual
Flag: legacy CUSTOM_LPNH_THRESHOLDS enabled
Change-Id: I3f10b944257e7021ae7c930b782f1d963c0e505b