Prevent menu items from showing on split tasks
This CL changes the TaskShortcutFactory for SCREENSHOT and MODAL so that the associated menu options ("Screenshot" and "Select") don't show up on split tasks. These actions are not fully supported on splitscreen tiles (they will only work on one of the two apps), so we disable them.

Bug: 327434215
Flag: N/A
Test: Manual
Change-Id: I121a9c28fa7570f10e13be2de6489182e3362cfc