Ensure that the recents start time is non-zero
It was only assigned when using gesture navigation.

When using non-gesture navigation, the calculation below in
ActivityMetricsLogger will be (uptime - 0) which could be several
hours or days.
 mSourceEventDelayMs = (int) (TimeUnit.NANOSECONDS.toMillis(
     launchingState.mStartUptimeNs) - sourceInfo.eventTimeMs);

Because the start time was set right after GestureState is created,
it can be simply set as the creation time of GestureState.
So non-gesture navigation can provide the timestamp when handling
OverviewCommandHelper.TYPE_TOGGLE.

Bug: 206872204
Test: Use 3-button navigation to enter recent.
      The value of LatencyTracker should not be unexpected large.
Change-Id: Ie661222822912e287d1ac295ccaf49e2086d909e