Settle without jumping at the end of the home transition.
We dynamically target the actual position of the icon during the
animation, instead of the final position. This way, whenever we catch up
we can swap to the actual icon seamlessly. If we don't do this, the
spring ends before the scale and the is a snap from the final position
to the currently scaling position.

Bug: 298089923
Flag: ACONFIG com.android.launcher3.enable_scaling_reveal_home_animation DISABLED
Test: see videos in b/298089923#comment9
Change-Id: I36f19416552d4981b3d9f62a6eb387a08432dd4f