Add TestActivity without any label for testing
This change adds an unlabeled activity to test the overview select action. The test evaluates the overview modal task state to verify that the UI displays correctly.
Using an activity without a label prevents flakiness from AiAi, which sometimes shows the activity title as selected and sometimes does not.

Bug: 332750488
Flag: NA
Test: OverviewTaskSelectImageTest
Change-Id: I6b63f1410bc466ac6711b0ca3d609d8f7a31d8c1