Go to NORMAL state before adding widget with add button
Updates add button behavior to go to NORMAL state before adding the
widget to the workspace. This has the effect of dismissing the All Apps
view if it is open.

This fixes an issue where the add button did not work if used from the
WidgetsBottomSheet while All Apps is open, particularly if a new screen
had to be created. See screencasts in gerrit for before/after.

Bug: 335906123
Test: manual, add widget with button from all apps > bottom sheet while
 there is no space on the launcher
Flag: ACONFIG com.android.launcher3.enable_widget_tap_to_add NEXTFOOD
Change-Id: I2fd961c09c473045088feba5d572c26a812e6619