Move the Done button in gesture navigation tutorial
To align with updated UX specs, this change moves the Done button in
each gesture's tutorial from the middle of the screen to the bottom
right.

Fix: 293285057
Test: Run through the gesture tutorial in all font/display sizes for
different form factors and check that the Done button after each step is
in the new location.
Flag: LEGACY ENABLE_NEW_GESTURE_NAV_TUTORIAL ENABLED

Change-Id: I2f2a7be0bf365727cff030f827e539add8533c18