Scale down the widget previews on similar scale as the spring load scale
Currently entire workspace scales down, but the original view sticks
to user's finger at same size as before. This makes the view look larger
obstructing the view of workspace behind it - specifically, when the
widgets are of larger size.

Additionally, with this scaling change, the preview images that aren't
of same aspect ratio don't break the transition to final widget as
before.

Bug: 338290585
Test: Unit test
Flag: None bugfix
Change-Id: I1d1818e60506898c7b28074f01cee91282032a94