Fix navigation buttons focus state
Two focus states in the same screen because taskbar and launcher are two processes. And we could not use keyboard to navigate to the nav button, agreed with UX that it should not have focus state. Make it not focusable to remove focus state as we are unable to focus on it by keyboard anyways. Tested that it would not affect TalkBack and SwitchAccess.

Only one focus in testing video: https://drive.google.com/file/d/18X3dtsvqfCMpdI7sBRxnaKwUCQzhHn8n/view?usp=sharing

Bug: 327386991
Test: Manual, navigate by keyboard in homepage and ensure there is no focus in the nav buttons
Flag: ACONFIG com.android.launcher3.enable_focus_outline Nextfood
Change-Id: I289ecb1784f8eb9b58bb36b28059ce4f367cf84e