Fix bubble translation when released after dragging
Translation for bubble after drag release was based on bubble bar translation. TranslationX was adjusted for bubble position. And translationY was reused from bubble bar.
Before ag/27284635 bubble bar translationY just copied over the input translation value. But the cl introduced a fixed translationY applied to bubble bar.
This was getting applied to the bubble as well. And resulted in bubble being out of its target position when releasing drag.

Bug: 330585402
Test: drag a bubble from bubble bar and release it, observe it animates
  to its correct position in the bar

Change-Id: I48f2a84b60a9cc6047e9841c254b4035212f3c0e