Fix state after dragging bubble to dismiss
After dragging a bubble in the bubble bar to dismiss it,
we remove it from the view and update the width of the bubble bar.

The issue here is that the drag state only gets cleared in
BubbleBarView after the dismiss animation is complete, so when
the width is updated, we don't reposition all the bubbles according
to the new width, and end up in a weird state.

The fix is to clear the drag state in onDragDismiss.

Flag: com.android.wm.shell.enable_bubble_bar
Fixes: 345482103
Test: manual
       - Add multiple bubbles to the bubble bar
       - Drag a bubble to dismiss it
       - Observe bubble bar updates correctly
       - Repeat with other bubbles
Change-Id: Ie94f8ace0f5660466686a261174b42c8d71b3e30